:recycle: (PC-5955) move tracking logic in specific hook
